Fix dependencies, changed appendment of data
I changed the datetime format for the dependences and swapped the temp and level data for 284217 which had been switched accidentally. I changed the code for appending new data by having the code remove the past 90 days of data and replace it with the new data from the LevelSender sqlite which insures all data is saved because sometimes the LevelSenders skip days of reporting.